feat: Add hash verification page
