+++ src/HashVerificationPage.css
@@ -1,0 +1,6 @@
+.hash-field {
+    width: 100%;
+    margin-bottom: 20px;
+    padding: 10px;
+    margin-right: 10px;
+}
